Split large implementation files by responsibility
