+++ Common.vcxproj.filters
@@ -272,8 +272,15 @@
     <ClCompile Include="AbstractAPOInfo.cpp" />
     <ClCompile Include="ConfigurationFileReader.cpp" />
     <ClCompile Include="DeviceAPOInfo.cpp" />
+    <ClCompile Include="DeviceAPOInfo.Install.cpp" />
+    <ClCompile Include="DeviceAPOInfo.Load.cpp" />
+    <ClCompile Include="DeviceAPOInfo.State.cpp" />
+    <ClCompile Include="DeviceAPOInfo.Uninstall.cpp" />
     <ClCompile Include="FilterConfiguration.cpp" />
     <ClCompile Include="FilterEngine.cpp" />
+    <ClCompile Include="FilterEngine.Configuration.cpp" />
+    <ClCompile Include="FilterEngine.Process.cpp" />
+    <ClCompile Include="FilterEngine.Runtime.cpp" />
     <ClCompile Include="IFilter.cpp" />
     <ClCompile Include="filters\loudnessCorrection\VolumeController.cpp">
       <Filter>filters\loudnessCorrection</Filter>
