+++ DeviceAPOInfo.cpp
@@ -1,4 +1,4 @@
-﻿/*
+/*
 	This file is part of EqualizerAPO, a system-wide equalizer.
 	Copyright (C) 2012  Jonas Thedering
 
@@ -178,616 +178,4 @@ bool DeviceAPOInfo::checkAPORegistration(bool fix)
 	}
 
 	return result;
-}
-
-bool DeviceAPOInfo::load(const wstring& deviceGuid, wstring defaultDeviceGuid)
-{
-	wstring keyPath;
-	if (RegistryHelper::keyExists(renderKeyPath L"\\" + deviceGuid))
-	{
-		keyPath = renderKeyPath L"\\" + deviceGuid;
-		input = false;
-	}
-	else
-	{
-		keyPath = captureKeyPath L"\\" + deviceGuid;
-		input = true;
-	}
-
-	unsigned long deviceState = RegistryHelper::readDWORDValue(keyPath, L"DeviceState");
-	if (deviceState & DEVICE_STATE_NOTPRESENT)
-		return false;
-	// somehow, disabled devices do not actually use DEVICE_STATE_DISABLED in registry but 0x10000000
-	disabled = deviceState & DEVICE_STATE_DISABLED || deviceState & 0x10000000;
-	unplugged = deviceState & DEVICE_STATE_UNPLUGGED;
-
-	this->deviceGuid = deviceGuid;
-
-	connectionName = RegistryHelper::readValue(keyPath + L"\\Properties", connectionValueName);
-	deviceName = RegistryHelper::readValue(keyPath + L"\\Properties", deviceValueName);
-
-	channelCount = 0;
-	sampleRate = 0;
-	channelMask = 0;
-	if (RegistryHelper::valueExists(keyPath + L"\\Properties", formatValueName))
-	{
-		std::vector<unsigned char> format = RegistryHelper::readBinaryValue(keyPath + L"\\Properties", formatValueName);
-		if (format.size() >= sizeof(WAVEFORMATEX) + 8)
-		{
-			WAVEFORMATEX* waveFormat = (WAVEFORMATEX*)&format[8];
-			channelCount = waveFormat->nChannels;
-			sampleRate = waveFormat->nSamplesPerSec;
-			if (waveFormat->wFormatTag == WAVE_FORMAT_EXTENSIBLE)
-			{
-				WAVEFORMATEXTENSIBLE* waveFormatExtensible = (WAVEFORMATEXTENSIBLE*)waveFormat;
-				channelMask = waveFormatExtensible->dwChannelMask;
-			}
-		}
-	}
-	if (channelMask == 0 && RegistryHelper::valueExists(keyPath + L"\\Properties", channelMaskValueName))
-		channelMask = RegistryHelper::readDWORDValue(keyPath + L"\\Properties", channelMaskValueName);
-
-	if (defaultDeviceGuid == L"")
-		defaultDeviceGuid = getDefaultDevice(input);
-
-	GUID guid1, guid2;
-	if (SUCCEEDED(CLSIDFromString(deviceGuid.c_str(), &guid1)) && SUCCEEDED(CLSIDFromString(defaultDeviceGuid.c_str(), &guid2)))
-		defaultDevice = (guid1 == guid2) != 0;
-	else
-		defaultDevice = false;
-
-	enhancementsDisabled = false;
-	if (RegistryHelper::keyExists(keyPath + L"\\FxProperties") && RegistryHelper::valueExists(keyPath + L"\\FxProperties", disableEnhancementsValueName))
-		enhancementsDisabled = RegistryHelper::readDWORDValue(keyPath + L"\\FxProperties", disableEnhancementsValueName) != 0;
-
-	installed = false;
-	currentInstallState.installMode = INSTALL_LFX_GFX;
-	version = L"0";
-	preMixChildGuid = L"";
-	postMixChildGuid = L"";
-	currentInstallState.installPreMix = true;
-	currentInstallState.installPostMix = !input;
-	currentInstallState.useOriginalAPOPreMix = true;
-	currentInstallState.useOriginalAPOPostMix = !input;
-	currentInstallState.allowSilentBufferModification = false;
-	currentInstallState.autoAdjust = true;
-
-	if (!RegistryHelper::keyExists(keyPath + L"\\FxProperties"))
-	{
-		for (int i = 0; i < allGuidValueNameCount; i++)
-			originalApoGuids[i] = APOGUID_NOKEY;
-	}
-	else
-	{
-		for (int i = 0; i < allGuidValueNameCount; i++)
-		{
-			if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", allGuidValueNames[i]))
-			{
-				wstring originalApoGuid = RegistryHelper::readValue(keyPath + L"\\FxProperties", allGuidValueNames[i]);
-				if (originalApoGuid == RegistryHelper::getGuidString(EQUALIZERAPO_PRE_MIX_GUID) || originalApoGuid == RegistryHelper::getGuidString(EQUALIZERAPO_POST_MIX_GUID))
-					originalApoGuid = APOGUID_NOVALUE;
-				originalApoGuids[i] = originalApoGuid;
-			}
-			else
-			{
-				originalApoGuids[i] = APOGUID_NOVALUE;
-			}
-		}
-
-		bool found = false;
-		bool foundAt[allGuidValueNameCount];
-		for (int i = 0; i < allGuidValueNameCount; i++)
-		{
-			foundAt[i] = false;
-
-			if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", allGuidValueNames[i]))
-			{
-				wstring apoGuidString = RegistryHelper::readValue(keyPath + L"\\FxProperties", allGuidValueNames[i]);
-
-				GUID apoGuid;
-				if (SUCCEEDED(CLSIDFromString(apoGuidString.c_str(), &apoGuid)))
-				{
-					if (apoGuid == EQUALIZERAPO_PRE_MIX_GUID || apoGuid == EQUALIZERAPO_POST_MIX_GUID)
-					{
-						foundAt[i] = true;
-						found = true;
-					}
-				}
-			}
-		}
-
-		if (found)
-		{
-			installed = true;
-
-			if (RegistryHelper::keyExists(childApoPath L"\\" + deviceGuid))
-			{
-				if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, versionValueName))
-				{
-					version = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, versionValueName);
-					if (version != installVersion)
-						throw RegistryException(L"Unsupported version of APO installation detected! Please uninstall newer Equalizer APO before using this version of Device Selector.");
-				}
-				else
-				{
-					version = L"1";
-				}
-
-				for (int i = 0; i < allGuidValueNameCount; i++)
-				{
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, allGuidValueNames[i]))
-						originalApoGuids[i] = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, allGuidValueNames[i]);
-				}
-
-				if (version == installVersion)
-				{
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, preMixChildGuidValueName))
-						preMixChildGuid = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, preMixChildGuidValueName);
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, postMixChildGuidValueName))
-						postMixChildGuid = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, postMixChildGuidValueName);
-
-					currentInstallState.installPreMix = foundAt[LFX_INDEX] || foundAt[SFX_INDEX];
-					currentInstallState.installPostMix = foundAt[GFX_INDEX] || foundAt[MFX_INDEX] || foundAt[EFX_INDEX];
-
-					if (preMixChildGuid == L"")
-						currentInstallState.useOriginalAPOPreMix = false;
-					if (postMixChildGuid == L"")
-						currentInstallState.useOriginalAPOPostMix = false;
-
-					if (foundAt[LFX_INDEX] || foundAt[GFX_INDEX])
-						currentInstallState.installMode = INSTALL_LFX_GFX;
-					else if (foundAt[EFX_INDEX])
-						currentInstallState.installMode = INSTALL_SFX_EFX;
-					else if (foundAt[SFX_INDEX] || foundAt[MFX_INDEX])
-						currentInstallState.installMode = INSTALL_SFX_MFX;
-
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, allowSilentBufferValueName))
-						currentInstallState.allowSilentBufferModification = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, allowSilentBufferValueName) != L"false";
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, disableAutoAdjustValueName))
-						currentInstallState.autoAdjust = RegistryHelper::readValue(childApoPath L"\\" + deviceGuid, disableAutoAdjustValueName) == L"false";
-				}
-				else
-				{
-					if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, lfxGuidValueName))
-					{
-						preMixChildGuid = originalApoGuids[0];
-						postMixChildGuid = originalApoGuids[1];
-					}
-					else
-					{
-						preMixChildGuid = originalApoGuids[2];
-						postMixChildGuid = originalApoGuids[3];
-					}
-				}
-			}
-			else if (RegistryHelper::keyExists(childApoPath) && RegistryHelper::valueExists(childApoPath, deviceGuid))
-			{
-				for (int i = 0; i < allGuidValueNameCount; i++)
-				{
-					if (foundAt[i])
-					{
-						originalApoGuids[i] = RegistryHelper::readValue(childApoPath, deviceGuid);
-						if (i == LFX_INDEX || i == SFX_INDEX)
-							preMixChildGuid = originalApoGuids[i];
-						else
-							postMixChildGuid = originalApoGuids[i];
-						break;
-					}
-				}
-			}
-		}
-		else
-		{
-			if (RegistryHelper::isWindowsVersionAtLeast(6, 3)) // Windows 8.1
-			{
-				// only use LFX/GFX if the audio driver supplied only those APOs
-				if (RegistryHelper::keyExists(keyPath + L"\\FxProperties")
-					&& (RegistryHelper::valueExists(keyPath + L"\\FxProperties", lfxGuidValueName) || RegistryHelper::valueExists(keyPath + L"\\FxProperties", gfxGuidValueName))
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", sfxGuidValueName)
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", mfxGuidValueName)
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", efxGuidValueName)
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", multiSfxGuidValueName)
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", multiMfxGuidValueName)
-					&& !RegistryHelper::valueExists(keyPath + L"\\FxProperties", multiEfxGuidValueName))
-					currentInstallState.installMode = INSTALL_LFX_GFX;
-				// bluetooth devices may be combined in Windows 11, EFX will not work then
-				else if (RegistryHelper::valueExists(keyPath + L"\\Properties", combinedDeviceValueName))
-					currentInstallState.installMode = INSTALL_SFX_MFX;
-				else
-					currentInstallState.installMode = INSTALL_SFX_EFX;
-			}
-		}
-	}
-
-	return true;
-}
-
-bool DeviceAPOInfo::canBeUpgraded() const
-{
-	return installed && version != installVersion;
-}
-
-bool DeviceAPOInfo::hasChanges() const
-{
-	return installed && selectedInstallState != currentInstallState;
-}
-
-bool DeviceAPOInfo::isExperimental() const
-{
-	return !installed && originalApoGuids[0] == APOGUID_NOKEY;
-}
-
-wstring DeviceAPOInfo::getOriginalAPOPreMix()
-{
-	wstring guid;
-	switch (selectedInstallState.installMode)
-	{
-	case INSTALL_LFX_GFX:
-		guid = originalApoGuids[LFX_INDEX];
-		if (RegistryHelper::isWindowsVersionAtLeast(6, 3)) // Windows 8.1
-		{
-			if (originalApoGuids[LFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[GFX_INDEX] == APOGUID_NOVALUE)
-				guid = originalApoGuids[SFX_INDEX];
-		}
-		break;
-	case INSTALL_SFX_MFX:
-		guid = originalApoGuids[SFX_INDEX];
-		if (originalApoGuids[SFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[MFX_INDEX] == APOGUID_NOVALUE)
-			guid = originalApoGuids[LFX_INDEX];
-		break;
-	case INSTALL_SFX_EFX:
-		guid = originalApoGuids[SFX_INDEX];
-		if (originalApoGuids[SFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[EFX_INDEX] == APOGUID_NOVALUE)
-			guid = originalApoGuids[LFX_INDEX];
-		break;
-	}
-
-	if (guid == APOGUID_NOKEY || guid == APOGUID_NOVALUE)
-		guid = L"";
-
-	return guid;
-}
-
-wstring DeviceAPOInfo::getOriginalAPOPostMix()
-{
-	wstring guid;
-	switch (selectedInstallState.installMode)
-	{
-	case INSTALL_LFX_GFX:
-		guid = originalApoGuids[GFX_INDEX];
-		if (RegistryHelper::isWindowsVersionAtLeast(6, 3)) // Windows 8.1
-		{
-			if (originalApoGuids[LFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[GFX_INDEX] == APOGUID_NOVALUE)
-				guid = originalApoGuids[MFX_INDEX];
-		}
-		break;
-	case INSTALL_SFX_MFX:
-		guid = originalApoGuids[MFX_INDEX];
-		if (originalApoGuids[SFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[MFX_INDEX] == APOGUID_NOVALUE)
-			guid = originalApoGuids[GFX_INDEX];
-		break;
-	case INSTALL_SFX_EFX:
-		guid = originalApoGuids[EFX_INDEX];
-		if (originalApoGuids[SFX_INDEX] == APOGUID_NOVALUE && originalApoGuids[EFX_INDEX] == APOGUID_NOVALUE)
-			guid = originalApoGuids[GFX_INDEX];
-		break;
-	}
-
-	if (guid == APOGUID_NOKEY || guid == APOGUID_NOVALUE)
-		guid = L"";
-
-	return guid;
-}
-
-void DeviceAPOInfo::install()
-{
-	if (!selectedInstallState.installPreMix && !selectedInstallState.installPostMix)
-		return;
-
-	RegistryHelper::createKey(childApoPath);
-	RegistryHelper::createKey(childApoPath L"\\" + deviceGuid);
-
-	wstring keyPath;
-	if (!input)
-		keyPath = renderKeyPath L"\\" + deviceGuid;
-	else
-		keyPath = captureKeyPath L"\\" + deviceGuid;
-
-	if (!RegistryHelper::keyExists(keyPath + L"\\FxProperties"))
-	{
-		try
-		{
-			RegistryHelper::createKey(keyPath + L"\\FxProperties");
-		}
-		catch (const RegistryException&)
-		{
-			// Permissions were not sufficient, so change them
-			RegistryHelper::takeOwnership(keyPath);
-			RegistryHelper::makeWritable(keyPath);
-
-			RegistryHelper::createKey(keyPath + L"\\FxProperties");
-		}
-
-		RegistryHelper::writeValue(keyPath + L"\\FxProperties", fxTitleValueName, L"Equalizer APO");
-
-		for (int i = 0; i < allGuidValueNameCount; i++)
-		{
-			RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, allGuidValueNames[i], APOGUID_NOKEY);
-		}
-	}
-	else
-	{
-		vector<wstring> valuenames;
-
-		for (int i = 0; i < allGuidValueNameCount; i++)
-		{
-			wstring apoGuidString = APOGUID_NOVALUE;
-			if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", allGuidValueNames[i]))
-			{
-				apoGuidString = RegistryHelper::readValue(keyPath + L"\\FxProperties", allGuidValueNames[i]);
-				valuenames.push_back(allGuidValueNames[i]);
-			}
-
-			RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, allGuidValueNames[i], apoGuidString);
-		}
-
-		if (!valuenames.empty())
-			RegistryHelper::saveToFile(keyPath + L"\\FxProperties", valuenames,
-				L"backup_" + StringHelper::replaceIllegalCharacters(deviceName) + L"_" + StringHelper::replaceIllegalCharacters(connectionName) + L".reg");
-	}
-
-	wstring preMixValue;
-	wstring postMixValue;
-	if (selectedInstallState.useOriginalAPOPreMix)
-		preMixValue = getOriginalAPOPreMix();
-	if (selectedInstallState.useOriginalAPOPostMix)
-		postMixValue = getOriginalAPOPostMix();
-	RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, preMixChildGuidValueName, preMixValue);
-	RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, postMixChildGuidValueName, postMixValue);
-
-	RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, allowSilentBufferValueName, selectedInstallState.allowSilentBufferModification ? L"true" : L"false");
-	if (selectedInstallState.autoAdjust)
-	{
-		if (RegistryHelper::valueExists(childApoPath L"\\" + deviceGuid, disableAutoAdjustValueName))
-			RegistryHelper::deleteValue(childApoPath L"\\" + deviceGuid, disableAutoAdjustValueName);
-	}
-	else
-	{
-		RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, disableAutoAdjustValueName, L"true");
-	}
-	RegistryHelper::writeValue(childApoPath L"\\" + deviceGuid, versionValueName, installVersion);
-
-	if (selectedInstallState.installMode == INSTALL_LFX_GFX)
-	{
-		if (selectedInstallState.installPreMix)
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", lfxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_PRE_MIX_GUID));
-		if (selectedInstallState.installPostMix && !input)
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", gfxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_POST_MIX_GUID));
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", sfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", sfxGuidValueName);
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", mfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", mfxGuidValueName);
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", efxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", efxGuidValueName);
-	}
-	else if (selectedInstallState.installMode == INSTALL_SFX_MFX)
-	{
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", lfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", lfxGuidValueName);
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", gfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", gfxGuidValueName);
-		if (selectedInstallState.installPreMix)
-		{
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", sfxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_PRE_MIX_GUID));
-			if (!RegistryHelper::valueExists(keyPath + L"\\FxProperties", sfxProcessingModesValueName))
-				RegistryHelper::writeMultiValue(keyPath + L"\\FxProperties", sfxProcessingModesValueName, defaultProcessingModeValue);
-		}
-		if (selectedInstallState.installPostMix && !input)
-		{
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", mfxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_POST_MIX_GUID));
-			if (!RegistryHelper::valueExists(keyPath + L"\\FxProperties", mfxProcessingModesValueName))
-				RegistryHelper::writeMultiValue(keyPath + L"\\FxProperties", mfxProcessingModesValueName, defaultProcessingModeValue);
-		}
-		// don't change efx
-	}
-	else if (selectedInstallState.installMode == INSTALL_SFX_EFX)
-	{
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", lfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", lfxGuidValueName);
-		if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", gfxGuidValueName))
-			RegistryHelper::deleteValue(keyPath + L"\\FxProperties", gfxGuidValueName);
-		if (selectedInstallState.installPreMix)
-		{
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", sfxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_PRE_MIX_GUID));
-			if (!RegistryHelper::valueExists(keyPath + L"\\FxProperties", sfxProcessingModesValueName))
-				RegistryHelper::writeMultiValue(keyPath + L"\\FxProperties", sfxProcessingModesValueName, defaultProcessingModeValue);
-		}
-		// don't change mfx
-		if (selectedInstallState.installPostMix && !input)
-		{
-			RegistryHelper::writeValue(keyPath + L"\\FxProperties", efxGuidValueName, RegistryHelper::getGuidString(EQUALIZERAPO_POST_MIX_GUID));
-			if (!RegistryHelper::valueExists(keyPath + L"\\FxProperties", efxProcessingModesValueName))
-				RegistryHelper::writeMultiValue(keyPath + L"\\FxProperties", efxProcessingModesValueName, defaultProcessingModeValue);
-		}
-	}
-
-	// force-enable enhancements
-	if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", disableEnhancementsValueName))
-		RegistryHelper::deleteValue(keyPath + L"\\FxProperties", disableEnhancementsValueName);
-}
-
-void DeviceAPOInfo::uninstall()
-{
-	wstring keyPath;
-	if (!input)
-		keyPath = renderKeyPath L"\\" + deviceGuid;
-	else
-		keyPath = captureKeyPath L"\\" + deviceGuid;
-
-	if (originalApoGuids[0] == APOGUID_NOKEY)
-	{
-		RegistryHelper::deleteKey(keyPath + L"\\FxProperties");
-	}
-	else
-	{
-		for (int i = 0; i < allGuidValueNameCount; i++)
-		{
-			if (originalApoGuids[i] == APOGUID_NOVALUE)
-			{
-				if (RegistryHelper::valueExists(keyPath + L"\\FxProperties", allGuidValueNames[i]))
-					RegistryHelper::deleteValue(keyPath + L"\\FxProperties", allGuidValueNames[i]);
-			}
-			else if (originalApoGuids[i] != L"")
-			{
-				RegistryHelper::writeValue(keyPath + L"\\FxProperties", allGuidValueNames[i], originalApoGuids[i]);
-			}
-		}
-	}
-
-	if (RegistryHelper::keyExists(childApoPath) && RegistryHelper::valueExists(childApoPath, deviceGuid))
-		RegistryHelper::deleteValue(childApoPath, deviceGuid);
-
-	if (RegistryHelper::keyExists(childApoPath L"\\" + deviceGuid))
-		RegistryHelper::deleteKey(childApoPath L"\\" + deviceGuid);
-
-	if (RegistryHelper::keyExists(childApoPath) && RegistryHelper::keyEmpty(childApoPath))
-		RegistryHelper::deleteKey(childApoPath);
-}
-
-void DeviceAPOInfo::reinstall()
-{
-	uninstall();
-	load(deviceGuid);
-	install();
-}
-
-wstring DeviceAPOInfo::getConnectionName() const
-{
-	return connectionName;
-}
-
-wstring DeviceAPOInfo::getDeviceName() const
-{
-	return deviceName;
-}
-
-wstring DeviceAPOInfo::getDeviceGuid() const
-{
-	return deviceGuid;
-}
-
-wstring DeviceAPOInfo::getDeviceString() const
-{
-	return getConnectionName() + L" " + getDeviceName() + L" " + getDeviceGuid();
-}
-
-unsigned DeviceAPOInfo::getChannelCount() const
-{
-	return channelCount;
-}
-
-unsigned DeviceAPOInfo::getSampleRate() const
-{
-	return sampleRate;
-}
-
-unsigned long DeviceAPOInfo::getChannelMask() const
-{
-	return channelMask;
-}
-
-bool DeviceAPOInfo::isInput() const
-{
-	return input;
-}
-
-bool DeviceAPOInfo::isInstalled() const
-{
-	return installed;
-}
-
-bool DeviceAPOInfo::isEnhancementsDisabled() const
-{
-	return enhancementsDisabled;
-}
-
-bool DeviceAPOInfo::isDefaultDevice() const
-{
-	return defaultDevice;
-}
-
-bool DeviceAPOInfo::isDisabled() const
-{
-	return disabled;
-}
-
-bool DeviceAPOInfo::isUnplugged() const
-{
-	return unplugged;
-}
-
-const DeviceAPOInfo::InstallState& DeviceAPOInfo::getCurrentInstallState()
-{
-	return currentInstallState;
-}
-
-DeviceAPOInfo::InstallState& DeviceAPOInfo::getSelectedInstallState()
-{
-	return selectedInstallState;
-}
-
-wstring DeviceAPOInfo::getPreMixChildGuid()
-{
-	return preMixChildGuid;
-}
-
-wstring DeviceAPOInfo::getPostMixChildGuid()
-{
-	return postMixChildGuid;
-}
-
-void DeviceAPOInfo::testAPOInstallation()
-{
-	IMMDeviceEnumerator* enumerator = nullptr;
-	IMMDevice* device = nullptr;
-	IAudioClient* audioClient = nullptr;
-	WAVEFORMATEX* format = nullptr;
-
-	HRESULT hr = CoCreateInstance(__uuidof(MMDeviceEnumerator), nullptr, CLSCTX_ALL, __uuidof(IMMDeviceEnumerator), (void**)&enumerator);
-	if (FAILED(hr))
-		fail(L"CoCreateInstance for IMMDeviceEnumerator", hr);
-	SCOPE_EXIT{enumerator->Release(); };
-
-	hr = enumerator->GetDevice(((input ? L"{0.0.1.00000000}." : L"{0.0.0.00000000}.") + deviceGuid).c_str(), &device);
-	if (FAILED(hr))
-		fail(L"GetDevice", hr);
-	SCOPE_EXIT{device->Release(); };
-
-	DWORD state;
-	hr = device->GetState(&state);
-	if (FAILED(hr))
-		fail(L"GetState", hr);
-	if (state & DEVICE_STATE_DISABLED || state & DEVICE_STATE_UNPLUGGED)
-		return;
-
-	hr = device->Activate(__uuidof(IAudioClient), CLSCTX_ALL, nullptr, (void**)&audioClient);
-	if (FAILED(hr))
-		fail(L"Activate", hr);
-	SCOPE_EXIT{audioClient->Release(); };
-
-	hr = audioClient->GetMixFormat(&format);
-	if (FAILED(hr))
-		fail(L"GetMixFormat", hr);
-	SCOPE_EXIT{CoTaskMemFree(format); };
-
-	hr = audioClient->Initialize(AUDCLNT_SHAREMODE_SHARED, 0, 1000000 /*100 ms*/, 0, format, nullptr);
-	if (FAILED(hr))
-		fail(L"Initialize", hr);
-}
-
-void DeviceAPOInfo::fail(const wstring& functionName, HRESULT hr)
-{
-	_com_error err(hr);
-	const wchar_t* msg = err.ErrorMessage();
-	throw DeviceException(functionName + L" failed for device \"" + deviceName + L"\" (" + msg + L")");
 }
