+++ Editor/Editor.pro
@@ -31,6 +31,11 @@ SOURCES += main.cpp\
 	guis/CommentFilterGUIFactory.cpp \
 	guis/CommentFilterGUI.cpp \
 	FilterTable.cpp \
+	FilterTable.Clipboard.cpp \
+	FilterTable.DragDrop.cpp \
+	FilterTable.Events.cpp \
+	FilterTable.Model.cpp \
+	FilterTable.Mouse.cpp \
 	../filters/PreampFilter.cpp \
 	../filters/PreampFilterFactory.cpp \
 	../helpers/MemoryHelper.cpp \
@@ -39,6 +44,10 @@ SOURCES += main.cpp\
 	guis/DeviceFilterGUI.cpp \
 	guis/DeviceFilterGUIFactory.cpp \
 	../DeviceAPOInfo.cpp \
+	../DeviceAPOInfo.Install.cpp \
+	../DeviceAPOInfo.Load.cpp \
+	../DeviceAPOInfo.State.cpp \
+	../DeviceAPOInfo.Uninstall.cpp \
 	guis/DeviceFilterGUIDialog.cpp \
 	../filters/DeviceFilterFactory.cpp \
 	guis/ChannelFilterGUIDialog.cpp \
@@ -94,6 +103,13 @@ SOURCES += main.cpp\
 	helpers/DisableWheelFilter.cpp \
 	widgets/EscapableLineEdit.cpp \
 	MainWindow.cpp \
+	MainWindow.Analysis.cpp \
+	MainWindow.Device.cpp \
+	MainWindow.Edit.cpp \
+	MainWindow.FileActions.cpp \
+	MainWindow.FileIO.cpp \
+	MainWindow.Preferences.cpp \
+	MainWindow.ViewActions.cpp \
 	guis/StageFilterGUI.cpp \
 	guis/StageFilterGUIFactory.cpp \
 	guis/ExpressionFilterGUIFactory.cpp \
@@ -101,6 +117,9 @@ SOURCES += main.cpp\
 	AnalysisPlotView.cpp \
 	AnalysisPlotScene.cpp \
 	../FilterEngine.cpp \
+	../FilterEngine.Configuration.cpp \
+	../FilterEngine.Process.cpp \
+	../FilterEngine.Runtime.cpp \
 	../FilterConfiguration.cpp \
 	../filters/ChannelFilterFactory.cpp \
 	../filters/ExpressionFilterFactory.cpp \
