+++ Editor/MainWindow.FileActions.cpp
@@ -1,0 +1,189 @@
+#include <sstream>
+#include <QDrag>
+#include <QElapsedTimer>
+#include <QLabel>
+#include <QMimeData>
+#include <QPushButton>
+#include <QStandardItemModel>
+#include <QStringBuilder>
+#include <QScrollArea>
+#include <QFileInfo>
+#include <QFileDialog>
+#include <QMessageBox>
+#include <QProcess>
+#include <QSettings>
+
+#define WIN32_LEAN_AND_MEAN
+#include <windows.h>
+#include <shellapi.h>
+
+#include "helpers/StringHelper.h"
+#include "helpers/LogHelper.h"
+#include "helpers/ChannelHelper.h"
+#include "Editor/helpers/GUIChannelHelper.h"
+#include "Editor/helpers/GUIHelper.h"
+#include "version.h"
+#include "FilterTable.h"
+#include "MainWindow.h"
+#include "ui_MainWindow.h"
+
+using std::find;
+using std::list;
+using std::set;
+using std::shared_ptr;
+using std::string;
+using std::stringstream;
+using std::vector;
+using std::wstring;
+
+
+void MainWindow::on_actionOpen_triggered()
+{
+	QScrollArea* scrollArea = qobject_cast<QScrollArea*>(ui->tabWidget->currentWidget());
+	QString path;
+	if (scrollArea != nullptr)
+	{
+		FilterTable* filterTable = qobject_cast<FilterTable*>(scrollArea->widget());
+		if (filterTable->getConfigPath().length() > 0)
+		{
+			QFileInfo fileInfo(filterTable->getConfigPath());
+			path = fileInfo.absolutePath();
+		}
+	}
+	if (path.length() == 0)
+		path = configDir.absolutePath();
+
+	QFileDialog dialog(this, tr("Open file"), path, "*.txt");
+	dialog.setFileMode(QFileDialog::ExistingFiles);
+	dialog.setNameFilter(tr("E-APO configurations (*.txt)"));
+
+	if (dialog.exec() == QDialog::Accepted)
+	{
+		for (QString file : dialog.selectedFiles())
+			load(file);
+	}
+}
+
+void MainWindow::on_actionSave_triggered()
+{
+	QScrollArea* scrollArea = qobject_cast<QScrollArea*>(ui->tabWidget->currentWidget());
+	if (scrollArea == nullptr)
+		return;
+
+	FilterTable* filterTable = qobject_cast<FilterTable*>(scrollArea->widget());
+
+	if (filterTable->getConfigPath().length() == 0)
+	{
+		ui->actionSaveAs->trigger();
+	}
+	else
+	{
+		save(filterTable, filterTable->getConfigPath());
+
+		QString tabText = ui->tabWidget->tabText(ui->tabWidget->currentIndex());
+		if (tabText.endsWith('*'))
+			ui->tabWidget->setTabText(ui->tabWidget->currentIndex(), tabText.left(tabText.length() - 1));
+	}
+}
+
+void MainWindow::on_actionSaveAs_triggered()
+{
+	QScrollArea* scrollArea = qobject_cast<QScrollArea*>(ui->tabWidget->currentWidget());
+	if (scrollArea == nullptr)
+		return;
+
+	FilterTable* filterTable = qobject_cast<FilterTable*>(scrollArea->widget());
+	QString path;
+	QString filename;
+	if (filterTable->getConfigPath().length() == 0)
+	{
+		path = configDir.absolutePath();
+	}
+	else
+	{
+		QFileInfo fileInfo(filterTable->getConfigPath());
+		path = fileInfo.absolutePath();
+		filename = fileInfo.fileName();
+	}
+	QFileDialog dialog(this, tr("Save file as"), path, "*.txt");
+	dialog.setFileMode(QFileDialog::AnyFile);
+	dialog.setAcceptMode(QFileDialog::AcceptSave);
+	dialog.setNameFilter(tr("E-APO configurations (*.txt)"));
+	dialog.setDefaultSuffix(".txt");
+	if (filename.length() > 0)
+		dialog.selectFile(filename);
+
+	if (dialog.exec() == QDialog::Accepted)
+	{
+		QString savePath = dialog.selectedFiles().at(0);
+		save(filterTable, savePath);
+		filterTable->setConfigPath(QDir::toNativeSeparators(savePath));
+
+		QFileInfo fileInfo(savePath);
+		ui->tabWidget->setTabText(ui->tabWidget->currentIndex(), fileInfo.fileName());
+		ui->tabWidget->setTabToolTip(ui->tabWidget->currentIndex(), QDir::toNativeSeparators(fileInfo.absoluteFilePath()));
+	}
+}
+
+void MainWindow::on_actionNew_triggered()
+{
+	FilterTable* filterTable = addTab(tr("Unsaved"), "", "", QList<QString>());
+
+	connect(filterTable, SIGNAL(linesChanged()), this, SLOT(linesChanged()));
+	ui->tabWidget->setCurrentIndex(ui->tabWidget->count() - 1);
+}
+
+void MainWindow::recentFileSelected()
+{
+	QAction* action = qobject_cast<QAction*>(sender());
+	load(action->text());
+}
+
+
+bool MainWindow::askForClose(int tabIndex)
+{
+	bool discarded = false;
+	if (ui->tabWidget->tabText(tabIndex).endsWith('*'))
+	{
+		ui->tabWidget->setCurrentIndex(tabIndex);
+		QString configPath = ui->tabWidget->tabToolTip(tabIndex);
+		QMessageBox messageBox;
+		messageBox.setWindowTitle(tr("Unsaved changes"));
+		messageBox.setText(tr("The configuration file %0 has unsaved changes.").arg(configPath));
+		messageBox.setInformativeText(tr("Do you want to save the changes before closing the file?"));
+		messageBox.setIcon(QMessageBox::Question);
+		messageBox.setStandardButtons(QMessageBox::Save | QMessageBox::Discard | QMessageBox::Cancel);
+		messageBox.setDefaultButton(QMessageBox::Save);
+		messageBox.setEscapeButton(QMessageBox::Cancel);
+		messageBox.setWindowFlags(Qt::Dialog | Qt::CustomizeWindowHint | Qt::WindowTitleHint | Qt::WindowCloseButtonHint);
+		int result = messageBox.exec();
+
+		switch (result)
+		{
+		case QMessageBox::Save:
+			ui->actionSave->trigger();
+			if (ui->tabWidget->tabText(tabIndex).endsWith('*'))
+				// saving was canceled
+				return false;
+			break;
+		case QMessageBox::Discard:
+			discarded = true;
+			break;
+		case QMessageBox::Cancel:
+			return false;
+		}
+	}
+
+	if (!discarded && !noSaveFilePreferences)
+	{
+		QScrollArea* scrollArea = qobject_cast<QScrollArea*>(ui->tabWidget->widget(tabIndex));
+		if (scrollArea != nullptr)
+		{
+			FilterTable* filterTable = qobject_cast<FilterTable*>(scrollArea->widget());
+			filterTable->savePreferences();
+		}
+	}
+
+	return true;
+}
+
