+++ Editor/MainWindow.ViewActions.cpp
@@ -1,0 +1,134 @@
+#include <sstream>
+#include <QDrag>
+#include <QElapsedTimer>
+#include <QLabel>
+#include <QMimeData>
+#include <QPushButton>
+#include <QStandardItemModel>
+#include <QStringBuilder>
+#include <QScrollArea>
+#include <QFileInfo>
+#include <QFileDialog>
+#include <QMessageBox>
+#include <QProcess>
+#include <QSettings>
+
+#define WIN32_LEAN_AND_MEAN
+#include <windows.h>
+#include <shellapi.h>
+
+#include "helpers/StringHelper.h"
+#include "helpers/LogHelper.h"
+#include "helpers/ChannelHelper.h"
+#include "Editor/helpers/GUIChannelHelper.h"
+#include "Editor/helpers/GUIHelper.h"
+#include "version.h"
+#include "FilterTable.h"
+#include "MainWindow.h"
+#include "ui_MainWindow.h"
+
+using std::find;
+using std::list;
+using std::set;
+using std::shared_ptr;
+using std::string;
+using std::stringstream;
+using std::vector;
+using std::wstring;
+
+
+void MainWindow::on_mainToolBar_visibilityChanged(bool visible)
+{
+	ui->actionToolbar->setChecked(visible);
+}
+
+void MainWindow::on_analysisDockWidget_visibilityChanged(bool visible)
+{
+	ui->actionAnalysisPanel->setChecked(visible);
+
+	if (visible)
+		startAnalysis();
+}
+
+void MainWindow::on_actionToolbar_triggered(bool checked)
+{
+	ui->mainToolBar->setVisible(checked);
+}
+
+void MainWindow::on_actionAnalysisPanel_triggered(bool checked)
+{
+	ui->analysisDockWidget->setVisible(checked);
+}
+
+void MainWindow::languageSelected(bool selected)
+{
+	QAction* action = qobject_cast<QAction*>(sender());
+
+	if (!selected)
+	{
+		action->setChecked(true);
+		return;
+	}
+
+	QLocale::Language language = static_cast<QLocale::Language>(action->data().toInt());
+
+	if (QMessageBox::question(this, tr("Restart required"), tr("Configuration Editor will be restarted to apply the changed settings. Proceed?")) == QMessageBox::Yes)
+	{
+		QSettings settings(QString::fromWCharArray(EDITOR_REGPATH), QSettings::NativeFormat);
+		if (language == QLocale::AnyLanguage)
+		{
+			settings.remove("language");
+		}
+		else
+		{
+			QString name = QLocale(language).name();
+			int index = name.indexOf('_');
+			if (index != -1)
+				name = name.left(index);
+			settings.setValue("language", name);
+		}
+
+		restart = true;
+		close();
+	}
+	else
+	{
+		action->setChecked(false);
+	}
+}
+
+void MainWindow::on_actionResetAllGlobalPreferences_triggered()
+{
+	if (QMessageBox::question(this, tr("Restart required"), tr("Configuration Editor will be restarted to apply the changed settings. Proceed?")) == QMessageBox::Yes)
+	{
+		QSettings settings(QString::fromWCharArray(EDITOR_REGPATH), QSettings::NativeFormat);
+		for (const QString& key : settings.childGroups())
+		{
+			if (key != "file-specific")
+				settings.remove(key);
+		}
+		for (const QString& key : settings.childKeys())
+			settings.remove(key);
+
+		restart = true;
+		noSavePreferences = true;
+		close();
+	}
+}
+
+void MainWindow::on_actionResetAllFileSpecificPreferences_triggered()
+{
+	if (QMessageBox::question(this, tr("Restart required"), tr("Configuration Editor will be restarted to apply the changed settings. Proceed?")) == QMessageBox::Yes)
+	{
+		QSettings settings(QString::fromWCharArray(EDITOR_PER_FILE_REGPATH), QSettings::NativeFormat);
+		for (const QString& key : settings.childGroups())
+			settings.remove(key);
+		for (const QString& key : settings.childKeys())
+			settings.remove(key);
+
+		restart = true;
+		noSaveFilePreferences = true;
+		close();
+	}
+}
+
